requirements.txt
requests==2.31.0
pytesseract==0.3.10
Pillow==10.1.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
